Seperate Versat API implementation from declaration
