+++ software/pc/versat.h
@@ -8,23 +8,23 @@
 
 //xconfdefs
 #define CONF_MEM0A 0
-#define CONF_ALU0 (CONF_MEM0A + 2*nMEM*MEMP_CONF_OFFSET)
-#define CONF_ALULITE0 (CONF_ALU0 + nALU*ALU_CONF_OFFSET)
-#define CONF_MUL0 (CONF_ALULITE0 + nALULITE*ALULITE_CONF_OFFSET)
-#define CONF_MULADD0 (CONF_MUL0 + nMUL*MUL_CONF_OFFSET)
-#define CONF_BS0 (CONF_MULADD0 + nMULADD*MULADD_CONF_OFFSET)
-#define CONF_REG_ADDR_W (clog2(CONF_BS0 + nBS*BS_CONF_OFFSET))
+#define CONF_ALU0 (CONF_MEM0A + 2 * nMEM * MEMP_CONF_OFFSET)
+#define CONF_ALULITE0 (CONF_ALU0 + nALU * ALU_CONF_OFFSET)
+#define CONF_MUL0 (CONF_ALULITE0 + nALULITE * ALULITE_CONF_OFFSET)
+#define CONF_MULADD0 (CONF_MUL0 + nMUL * MUL_CONF_OFFSET)
+#define CONF_BS0 (CONF_MULADD0 + nMULADD * MULADD_CONF_OFFSET)
+#define CONF_REG_ADDR_W (clog2(CONF_BS0 + nBS * BS_CONF_OFFSET))
 #define CONF_MEM_ADDR_W 3
-#define CONF_CLEAR (1<<CONF_REG_ADDR_W)
-#define GLOBAL_CONF_CLEAR (CONF_CLEAR+1)
-#define CONF_MEM (CONF_CLEAR + (1<<(CONF_REG_ADDR_W-1)))
+#define CONF_CLEAR (1 << CONF_REG_ADDR_W)
+#define GLOBAL_CONF_CLEAR (CONF_CLEAR + 1)
+#define CONF_MEM (CONF_CLEAR + (1 << (CONF_REG_ADDR_W - 1)))
 
 //xdefs
 #define nSTAGE_W (clog2(nSTAGE))
 #define nMEM_W (clog2(nMEM))
-#define CTR_ADDR_W (nSTAGE_W+2+nMEM_W+MEM_ADDR_W)
-#define N (2*nMEM + nALU + nALULITE + nMUL + nMULADD + nBS)
-#define N_W (clog2(N)+1)
+#define CTR_ADDR_W (nSTAGE_W + 2 + nMEM_W + MEM_ADDR_W)
+#define N (2 * nMEM + nALU + nALULITE + nMUL + nMULADD + nBS)
+#define N_W (clog2(N) + 1)
 
 //xmemdefs
 #define MEMP_CONF_ITER 0
@@ -128,6 +128,7 @@
 #define nMUL 2
 #define nMULADD 2
 #define nBS 1
+#define nMULADDLITE 0
 #define MEM_ADDR_W 5
 #define CONF_MEM_ADDR_W 3
